src/vogue_core/thumbnails.py
"""
Thumbnail generation utilities for Vogue Manager

Provides thumbnail generation for Maya files and other assets.
"""

import os
from pathlib import Path
from typing import Optional
from .logging_utils import get_logger


def make_thumbnail(source_scene_path: str, out_thumb_path: str, width: int = 256, height: int = 256) -> bool:
    """
    Generate a thumbnail for a scene file
    
    Args:
        source_scene_path: Path to source scene file
        out_thumb_path: Path for output thumbnail
        width: Thumbnail width
        height: Thumbnail height
        
    Returns:
        True if successful, False otherwise
    """
    logger = get_logger("Thumbnails")
    
    source_path = Path(source_scene_path)
    out_path = Path(out_thumb_path)
    
    if not source_path.exists():
        logger.error(f"Source file does not exist: {source_scene_path}")
        return False
    
    # Ensure output directory exists
    out_path.parent.mkdir(parents=True, exist_ok=True)
    
    # For now, create a placeholder thumbnail
    # TODO: Implement actual thumbnail generation
    # - Maya playblast integration
    # - Image extraction from scene files
    # - Support for different file formats
    
    try:
        # Create a simple placeholder image
        create_placeholder_thumbnail(str(out_path), width, height)
        logger.info(f"Created thumbnail: {out_thumb_path}")
        return True
    except Exception as e:
        logger.error(f"Failed to create thumbnail: {e}")
        return False


def create_placeholder_thumbnail(out_path: str, width: int = 256, height: int = 256) -> None:
    """
    Create a placeholder thumbnail image
    
    Args:
        out_path: Output file path
        width: Image width
        height: Image height
    """
    try:
        from PIL import Image, ImageDraw, ImageFont
        
        # Create image with dark background
        img = Image.new('RGB', (width, height), color='#2E2E2E')
        draw = ImageDraw.Draw(img)
        
        # Add border
        draw.rectangle([0, 0, width-1, height-1], outline='#00A1F1', width=2)
        
        # Add text
        try:
            # Try to use a default font
            font = ImageFont.load_default()
        except:
            font = None
        
        text = "Vogue\nManager"
        bbox = draw.textbbox((0, 0), text, font=font)
        text_width = bbox[2] - bbox[0]
        text_height = bbox[3] - bbox[1]
        
        x = (width - text_width) // 2
        y = (height - text_height) // 2
        
        draw.text((x, y), text, fill='#DCDCDC', font=font, align='center')
        
        # Save image
        img.save(out_path, 'JPEG', quality=85)
        
    except ImportError:
        # Fallback: create a simple text file
        with open(out_path.replace('.jpg', '.txt'), 'w') as f:
            f.write(f"Thumbnail placeholder for {Path(out_path).stem}\n")
            f.write(f"Size: {width}x{height}\n")
    except Exception as e:
        raise Exception(f"Failed to create placeholder thumbnail: {e}")


def get_thumbnail_path(version_path: str) -> str:
    """
    Get the expected thumbnail path for a version
    
    Args:
        version_path: Path to version file
        
    Returns:
        Expected thumbnail path
    """
    version_path = Path(version_path)
    
    # Create thumbnail in same directory with .jpg extension
    thumb_path = version_path.with_suffix('.jpg')
    
    # Alternative: create in dedicated thumbnails directory
    # thumb_dir = version_path.parent.parent.parent / "thumbs"
    # thumb_dir.mkdir(exist_ok=True)
    # thumb_path = thumb_dir / f"{version_path.stem}.jpg"
    
    return str(thumb_path)


def generate_thumbnail_for_version(version_path: str) -> Optional[str]:
    """
    Generate thumbnail for a version file
    
    Args:
        version_path: Path to version file
        
    Returns:
        Thumbnail path if successful, None otherwise
    """
    thumb_path = get_thumbnail_path(version_path)
    
    if make_thumbnail(version_path, thumb_path):
        return thumb_path
    else:
        return None


def batch_generate_thumbnails(version_paths: list) -> dict:
    """
    Generate thumbnails for multiple version files
    
    Args:
        version_paths: List of version file paths
        
    Returns:
        Dictionary mapping version paths to thumbnail paths
    """
    results = {}
    
    for version_path in version_paths:
        thumb_path = generate_thumbnail_for_version(version_path)
        results[version_path] = thumb_path
    
    return results
