pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vogue-manager"
version = "1.0.0"
description = "A Prism-like production pipeline management system for VFX and animation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Vogue Manager Team", email = "team@voguemanager.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "PySide2>=5.15.0",
    "pydantic>=1.8.0",
    "jsonschema>=3.2.0",
    "Pillow>=8.0.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0",
    "pytest-mock>=3.6.0",
    "pytest-qt>=4.0.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
    "isort>=5.9.0",
    "pre-commit>=2.15.0",
    "tox>=3.24.0",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=0.5.0",
]

[project.scripts]
vogue-manager = "vogue_app.main:main"
vogue-api = "vogue_core.api_local:run_api_server"

[project.urls]
Homepage = "https://github.com/voguemanager/vogue-manager"
Documentation = "https://voguemanager.readthedocs.io"
Repository = "https://github.com/voguemanager/vogue-manager.git"
Issues = "https://github.com/voguemanager/vogue-manager/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"vogue_maya" = ["*.mel"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/vogue_core",
    "--cov=src/vogue_app",
    "--cov=src/vogue_maya",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "gui: marks tests as GUI tests",
]
